hvf: fix licensing issues; isolate task handling code (GPL v2-only)
This patch replaces the license header for those files that were either
GPL v2-or-v3, or GPL v2-only; the replacing license is GPL v2-or-later.
The code for task switching/handling, which is derived from KVM and
hence is GPL v2-only, is isolated in the new files (with this license)
x86_task.c/.h, and the corresponding compilation rule is added to
target/i386/hvf-utils/Makefile.objs.
Note that rules for compiling are still missing from
target/i386/Makefile.objs; they will be added later in this patchset.

Signed-off-by: Sergio Andres Gomez Del Real <Sergio.G.DelReal@gmail.com>